Move SASL processing into a dedicated file.
Clean up some constants.  Add some behavior to response processing for
noops that allows it to handle (and swallow) errors that may occur in
a multi block
